Moved AuthConfig class to own file
